+++ hyperdbg/hprdbghv/header/common/State.h
@@ -304,7 +304,7 @@ typedef struct _VIRTUAL_MACHINE_STATE
     UINT64       TestNumber;                                                    // Used for test purposes (Number)
     GUEST_REGS * Regs;                                                          // The virtual processor's general-purpose registers
     UINT32       CoreId;                                                        // The core's unique identifier
-    ULONG        ExitReason;                                                    // The core's exit reason
+    UINT32        ExitReason;                                                    // The core's exit reason
     UINT32       ExitQualification;                                             // The core's exit qualification
     UINT64       LastVmexitRip;                                                 // RIP in the current VM-exit
     UINT64       VmxonRegionPhysicalAddress;                                    // Vmxon region physical address
